--- annotation/src/main/java/org/cbioportal/genome_nexus/annotation/domain/AnnotationEnricher.java
@@ -1,9 +1,0 @@
-package org.cbioportal.genome_nexus.annotation.domain;
-
-/**
- * @author Selcuk Onur Sumer
- */
-public interface AnnotationEnricher
-{
-    void enrich(VariantAnnotation annotation);
-}
