+++ annotation/src/main/java/org/cbioportal/genome_nexus/annotation/service/AnnotationEnricher.java
@@ -1,0 +1,11 @@
+package org.cbioportal.genome_nexus.annotation.service;
+
+import org.cbioportal.genome_nexus.annotation.domain.VariantAnnotation;
+
+/**
+ * @author Selcuk Onur Sumer
+ */
+public interface AnnotationEnricher
+{
+    void enrich(VariantAnnotation annotation);
+}
